+++ lib/src/SenseClient.pbjson.dart
@@ -5,33 +5,32 @@
 // @dart = 2.3
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
-const Request$json = const {
+const Request$json = {
   '1': 'Request',
-  '2': const [
-    const {'1': 'data', '3': 1, '4': 1, '5': 12, '10': 'data'},
-    const {'1': 'apikey', '3': 2, '4': 1, '5': 9, '10': 'apikey'},
-    const {'1': 'format', '3': 3, '4': 1, '5': 9, '10': 'format'},
-    const {'1': 'api_version', '3': 4, '4': 1, '5': 9, '10': 'apiVersion'},
-    const {'1': 'user_agent', '3': 5, '4': 1, '5': 9, '10': 'userAgent'},
+  '2': [
+    {'1': 'data', '3': 1, '4': 1, '5': 12, '10': 'data'},
+    {'1': 'apikey', '3': 2, '4': 1, '5': 9, '10': 'apikey'},
+    {'1': 'format', '3': 3, '4': 1, '5': 9, '10': 'format'},
+    {'1': 'api_version', '3': 4, '4': 1, '5': 9, '10': 'apiVersion'},
+    {'1': 'user_agent', '3': 5, '4': 1, '5': 9, '10': 'userAgent'},
   ],
 };
 
-const RequestStream$json = const {
+const RequestStream$json = {
   '1': 'RequestStream',
-  '2': const [
-    const {'1': 'data', '3': 1, '4': 1, '5': 12, '10': 'data'},
-    const {'1': 'apikey', '3': 2, '4': 1, '5': 9, '10': 'apikey'},
-    const {'1': 'dtype', '3': 3, '4': 1, '5': 9, '10': 'dtype'},
-    const {'1': 'sr', '3': 4, '4': 1, '5': 5, '10': 'sr'},
-    const {'1': 'api_version', '3': 5, '4': 1, '5': 9, '10': 'apiVersion'},
-    const {'1': 'user_agent', '3': 6, '4': 1, '5': 9, '10': 'userAgent'},
+  '2': [
+    {'1': 'data', '3': 1, '4': 1, '5': 12, '10': 'data'},
+    {'1': 'apikey', '3': 2, '4': 1, '5': 9, '10': 'apikey'},
+    {'1': 'dtype', '3': 3, '4': 1, '5': 9, '10': 'dtype'},
+    {'1': 'sr', '3': 4, '4': 1, '5': 5, '10': 'sr'},
+    {'1': 'api_version', '3': 5, '4': 1, '5': 9, '10': 'apiVersion'},
+    {'1': 'user_agent', '3': 6, '4': 1, '5': 9, '10': 'userAgent'},
   ],
 };
 
-const Response$json = const {
+const Response$json = {
   '1': 'Response',
-  '2': const [
-    const {'1': 'outputs', '3': 1, '4': 1, '5': 9, '10': 'outputs'},
+  '2': [
+    {'1': 'outputs', '3': 1, '4': 1, '5': 9, '10': 'outputs'},
   ],
 };
-