+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ﻿
+# 0.2.1
+Delete unnecessary code.
+
 # 0.2.0
-Update corresponding to Sense Client v1.0.0
+Update corresponding to Sense Client v1.0.0.
 
 # 0.1.0
 Add Doc Example.